+++ src/app/components/UserInfoMenu/index.tsx
@@ -431,10 +431,10 @@ class UserInfoMenu extends React.Component<IProps, IState> {
             settings.setMuteuntil(mode);
         } else if (mode > 0) {
             mode += this.riverTime.now();
+            settings.setMuteuntil(mode);
         }
         settings.setFlags(0);
         settings.setSound('');
-        settings.setMuteuntil(mode);
         this.sdk.setNotifySettings(peer, settings).then(() => {
             dialog.notifysettings = settings.toObject();
             this.setState({